requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-celery==3.2.2
idna==2.6
kombu==3.0.37
lxml==4.1.1
olefile==0.44
Pillow==4.3.0
pytesseract==0.1.7
pytz==2017.3
redis==2.10.6
requests==2.18.4
urllib3==1.22
